+++ src/element/textElement.ts
@@ -864,6 +864,7 @@ const VALID_CONTAINER_TYPES = new Set([
   "diamond",
   "image",
   "arrow",
+  "iframe",
 ]);
 
 export const isValidTextContainer = (element: ExcalidrawElement) =>