+++ src/element/typeChecks.ts
@@ -4,6 +4,7 @@ import { MarkNonNullable } from "../utility-types";
 import {
   ExcalidrawElement,
   ExcalidrawTextElement,
+  ExcalidrawIFrameElement,
   ExcalidrawLinearElement,
   ExcalidrawBindableElement,
   ExcalidrawGenericElement,
@@ -24,7 +25,8 @@ export const isGenericElement = (
     (element.type === "selection" ||
       element.type === "rectangle" ||
       element.type === "diamond" ||
-      element.type === "ellipse")
+      element.type === "ellipse" ||
+      element.type === "iframe")
   );
 };
 
@@ -38,6 +40,12 @@ export const isImageElement = (
   element: ExcalidrawElement | null,
 ): element is ExcalidrawImageElement => {
   return !!element && element.type === "image";
+};
+
+export const isIFrameElement = (
+  element: ExcalidrawElement | null,
+): element is ExcalidrawIFrameElement => {
+  return !!element && element.type === "iframe";
 };
 
 export const isTextElement = (
@@ -77,7 +85,7 @@ export const isArrowElement = (
 };
 
 export const isLinearElementType = (
-  elementType: AppState["activeTool"]["type"],
+  elementType: AppState["activeTool"]["type"] | "iframe",
 ): boolean => {
   return (
     elementType === "arrow" || elementType === "line" // || elementType === "freedraw"
@@ -96,7 +104,7 @@ export const isBindingElement = (
 };
 
 export const isBindingElementType = (
-  elementType: AppState["activeTool"]["type"],
+  elementType: AppState["activeTool"]["type"] | "iframe",
 ): boolean => {
   return elementType === "arrow";
 };
@@ -112,6 +120,7 @@ export const isBindableElement = (
       element.type === "diamond" ||
       element.type === "ellipse" ||
       element.type === "image" ||
+      element.type === "iframe" ||
       (element.type === "text" && !element.containerId))
   );
 };
@@ -127,6 +136,7 @@ export const isTextBindableContainer = (
       element.type === "diamond" ||
       element.type === "ellipse" ||
       element.type === "image" ||
+      element.type === "iframe" ||
       isArrowElement(element))
   );
 };
@@ -136,6 +146,7 @@ export const isExcalidrawElement = (element: any): boolean => {
     element?.type === "text" ||
     element?.type === "diamond" ||
     element?.type === "rectangle" ||
+    element?.type === "iframe" ||
     element?.type === "ellipse" ||
     element?.type === "arrow" ||
     element?.type === "freedraw" ||
@@ -163,7 +174,8 @@ export const isBoundToContainer = (
   );
 };
 
-export const isUsingAdaptiveRadius = (type: string) => type === "rectangle";
+export const isUsingAdaptiveRadius = (type: string) =>
+  type === "rectangle" || type === "iframe";
 
 export const isUsingProportionalRadius = (type: string) =>
   type === "line" || type === "arrow" || type === "diamond";
@@ -204,7 +216,7 @@ export const getDefaultRoundnessTypeForElement = (
     };
   }
 
-  if (element.type === "rectangle") {
+  if (element.type === "rectangle" || element.type === "iframe") {
     return {
       type: ROUNDNESS.ADAPTIVE_RADIUS,
     };