+++ src/packages/excalidraw/index.tsx
@@ -42,6 +42,8 @@ const ExcalidrawBase = (props: ExcalidrawProps) => {
     onPointerDown,
     onScrollChange,
     children,
+    iframeURLWhitelist,
+    renderCustomIFrame,
   } = props;
 
   const canvasActions = props.UIOptions?.canvasActions;
@@ -115,6 +117,8 @@ const ExcalidrawBase = (props: ExcalidrawProps) => {
           onLinkOpen={onLinkOpen}
           onPointerDown={onPointerDown}
           onScrollChange={onScrollChange}
+          iframeURLWhitelist={iframeURLWhitelist}
+          renderCustomIFrame={renderCustomIFrame}
         >
           {children}
         </App>